// src/enums/EColumnType.ts
export enum EColumnType {
  SKU = "sku",
  CATEGORY = "category",
  SUB_CATEGORY = "subCategory",
  DESCRIPTION = "description",
  PROVIDER = "provider",
  COUNTRY = "country",
  PACKING = "packing",
  BPP = "bpp",
  FORECAST = "forecast",
  CPT = "cpt",
  CIP = "cip",
  TABLE = "table",
  LT = "lt",
  FORECAST_ERROR = "forecastError",
  QTY_SOLD = "qtySold",
  P_SOLD = "pSold",
  PRICE = "price",
  COST = "cost",
  S_FACTOR = "sFactor",
  UTILITY_MARGIN = "utilityMargin",
  TOTAL_SALES = "totalSales",
  COST_SALES = "costSales",
  GROSS_MARGIN = "grossMargin",
  IPB = "ipb",
  AVG_INV = "avgInv",
  TRANSIT_INV = "transitInv",
  CURRENT_INV = "currentInv",
  ROTATION = "rotacion",
  MESES_INV = "mesesInv",
  PCK_SENT = "pckSent",
  NIVEL_SERVICIO_ACTUAL = "nivelServicioActual",
}